heightmap.vert: need higher precision to compute shadow

+++ lib/heightmap/heightmap/render/heightmap.vert
@@ -12,8 +12,8 @@ uniform mediump vec3 logScale;
 uniform mediump vec2 scale_tex;
 uniform mediump vec2 offset_tex;
 uniform highp mat4 ModelViewProjectionMatrix;
-uniform mediump mat4 ModelViewMatrix;
-uniform mediump mat4 NormalMatrix;
+uniform highp mat4 ModelViewMatrix;
+uniform highp mat4 NormalMatrix;
 
 mediump float heightValue(mediump float v) {
     v *= 0.01; // map from small range of float16 to float32 (TfrBlockUpdater)
@@ -50,29 +50,29 @@ void main()
     heighty1 = heightValue(heighty1);
     heighty2 = heightValue(heighty2);
 
-    mediump vec4 pos         = vec4(vertex.x, height, vertex.y, 1.0);
+    highp vec4 pos         = vec4(vertex.x, height, vertex.y, 1.0);
 
     // transform to homogeneous clip space
     gl_Position      = ModelViewProjectionMatrix * pos;
     vertex_height = height;
 
-    mediump vec2 slope       = vec2(heightx2-heightx1, heighty2-heighty1);
+    highp vec2 slope       = vec2(heightx2-heightx1, heighty2-heighty1);
 
     // calculate surface normal from slope for shading
-    mediump vec4 worldSpaceNormal;
+    highp vec4 worldSpaceNormal;
     worldSpaceNormal.xyz = cross( vec3(0.0,            slope.y, tex2.y-tex1.y),
                                    vec3(tex2.x-tex1.x,  slope.x, 0.0));
     worldSpaceNormal.w = 1.0;
 
-    mediump vec3 eyeSpacePos      = (ModelViewMatrix * pos).xyz;
-    mediump vec3 eyeSpaceNormal   = (NormalMatrix * worldSpaceNormal).xyz;
+    highp vec3 eyeSpacePos      = (ModelViewMatrix * pos).xyz;
+    highp vec3 eyeSpaceNormal   = (NormalMatrix * worldSpaceNormal).xyz;
 
     eyeSpacePos      = normalize(eyeSpacePos);
     eyeSpaceNormal   = normalize(eyeSpaceNormal);
     worldSpaceNormal = normalize(worldSpaceNormal);
 
-    mediump float facing    = max(0.0, dot(eyeSpaceNormal, -eyeSpacePos));
-    mediump float diffuse   = max(0.0, worldSpaceNormal.y); // max(0.0, dot(worldSpaceNormalVector, lightDir));
+    highp float facing    = max(0.0, dot(eyeSpaceNormal, -eyeSpacePos));
+    highp float diffuse   = max(0.0, worldSpaceNormal.y); // max(0.0, dot(worldSpaceNormalVector, lightDir));
 
     //shadow = clamp( 0.5 + diffuse+facing + fresnel, 0.5, 1.0);
     shadow = mix(1.0, min( 0.5 + (diffuse+facing)*0.5, 1.0), flatness);
